add wait option
Previously the script would wait for Sunday results since those are not
usually posted until Monday. This was useful for running the script on
Mondays. However, if those results are never posted, this can cause the
script to wait indefinitely - especially given the -r option.

To fix this, I added a -w option. The script will only wait for the
previous day's results if this -w option is specified. It doesn't make
any sense to wait for anything prior to yesterday as the posted PDF has
likely been updated and we'll never be able to recover the missing file.